Remove useless cast to usize

+++ flexo/src/main.rs
@@ -909,7 +909,7 @@ fn send_payload<T>(source: &mut File, filesize: u64, bytes_sent: i64, receiver: 
     let size = unsafe {
         let mut offset = bytes_sent as off64_t;
         while (offset as u64) < filesize {
-            let count = cmp::min(filesize as usize - offset as usize, MAX_SENDFILE_COUNT as usize);
+            let count = cmp::min(filesize as usize - offset as usize, MAX_SENDFILE_COUNT);
             debug!("Sendfile count: {}", count);
             let size: isize = libc::sendfile64(sfd, fd, &mut offset, count);
             if size == -1 {
